requirements.txt
torch==2.5.1
torchvision
torchaudio
fastapi
uvicorn
python-multipart
Pillow
numpy
opencv-python
einops
decord
pycocotools
python-doctr[torch]
shapely
easyocr
paddlepaddle-gpu==2.6.1
paddleocr>=2.7.0
pydantic-settings
python-dotenv
